WIP deployment updater refactor
